Intercept sigquit on ExUnit
By using `Ctrl+\ ` to shutdown the VM, ExUnit will
intercept the sigquit signal and print the tests
currently running. This can be used to see which
test is stuck in a long running test suite:

    $ mix test
    .^\

    Aborting test suite, the following have not completed:

    * test greets the world [test/example_test.exs:5]
    * AnotherTest [test/another_test.exs]

    Showing results so far...

    1 doctest, 0 failures

    Randomized with seed 825614

As shown above, we also include a summary of the run
so far, alongside the seed and failure reports.