Use defaults when the mix env or target are empty
If the user sets `$MIX_ENV` or `$MIX_TARGET` to the empty string rather
that unsetting the environment variables, `:""` will be passed for the
environment and target respectively. This will likely fail. For the mix
target, this mistake turned out to be hard to debug for new Nerves users
since mix target is much less used in other contexts.

This commit catches the empty string being passed to `$MIX_ENV` and
`$MIX_TARGET` and sets the environment and target to the defaults just
like if those variables were unset.