Preserve file mode when copying non-beam ebin files (#10709)
If a non-beam file in the ebin directory had execute permissions set,
those were lost when making a release. This switches the copy function
to `File.cp!/3` to preserve permissions.

Specifically, this affects the use of zigler with Nerves. Zigler
generates NIF shared libaries and stores them in the `ebin` directory.
The shared libraries had their execute mode bit set, but lost it on the
copy to the release. This affected the Nerves post-processing of all
firmware executables since Nerves uses the execute bit to identify
files that should be processed.

This also switches the copy call to the raising version like is done
when writing `.beam` files for consistency.