Cleanup waiting map on Task.Supervised.stream/6
When running Task.Supervised.stream/6 with ordered options turned off,
the "waiting" map to store the response from task is not cleaned up.
This leads to the response piled up in the map causing memory leak.

This changes cleanup the map of the response that has been delivered to the reducer
function.