Shorten is_struct/2 Macro Expansion for Atom Literals
The macro expansion for is_struct/2 includes a case that checks for
whether the second argument ("name") is an atom, and raises
ArgumentError if not. This is sometimes, maybe even usually, known at
compile time, and the second clause of the generated case is simply
dead code. Dialyzer warns about this.

The change is simply to check to see whether "name" is an atom at
compile-time, and if so not to generate the case at all but simply
perform the map __struct__ key check.