feat(mix loadconfig): reconfigure Erlang
This will reconfigure Erlang logger on `mix loadconfig` to provide
unified way to configure early-startup logging in development and
production environments.

It can cause some confusion for some users, that would try to configure
other `kernel` options in `config/config.ex`, however I think that these
cases are rare and are limited to the power-users, who already know the
limitations of this approach.

Of course there is an issue, that in the future Erlang can expand
supported configuration options that are available in `kernel` (I think
that the nearest thing that can be done there are global metadata).
However in my humble opinion it is price worth paying, as it will make
the logger configuration more straightforward and more integrated in
BEAM community as a whole.