Clarify `Path.extname/1` doc(test)s (#11023)
Documentation update for #11023 

Clarify the behavior of `Path.extname/1` in docstrings and doctests, namely how it behaves when given a filename that is just a file extension (like ".gitignore"). I also mentioned when you might want to use `Path.basename/1` instead, which was brought up by @josevalim in the associated issue (#11023).