Improve the error message when apply fails
Not all elixir developers may understand what "apply" means in this
case.  Especially if they are beginners and are not using
`Kernel.apply/3` directly (but instead using `.` to call the function or
if this is being raised by library code).

However, calling `:foo` a function (example from the tests) isn't quite
right (as it is an atom that could represent a function when paired with
a module) so the wording could probably be improved further.