Support sparse patterns in Mix.SCM.Git
`git sparse-checkout` supports the same pattern matching as `.gitignore`
files (see [this
article](https://github.blog/2020-01-17-bring-your-monorepo-down-to-size-with-sparse-checkout/#cone-mode-restricted-patterns-improve-performance)
for details). However, the current implementation of the `:sparse` option with
mix breaks if you try to use a pattern.

This is because `:sparse` is being added to `:dest` option which later
on attempts to be created via a `touch`, but patterns are not supported.

So this instead removes the update of the `:dest` option and uses the
`:sparse` option only when needed. As a bonus, this also adds support
for including multiple patterns as a list in `:sparse` to create more
complex sparse checkouts as needed.