Keep proper side of compile time transitivity
Imagine the following files:

    a -> b -> c (compile) -> d -> e

Previously when using the mix xref --label compile,
which is transitive, we were showing:

    a -> b -> c (compile)

However, a _will not_ recompile if c changes. The
correct is to show:

    b -> c (compile) -> d -> e

As b will recompile if any of c, d, or e changes.
This changes the printed graph into take this into
account correctly.