Add Atom.classify/1 and inspection functions
Today the formatter still uses some private APIs
and this commit aims to expose some of them.
In particular, Atom.classify/1 was added to expose
if an atom is an alias, an identifier, quoted, or
unquoted.

Note, however, there is no API for operators yet.
That's because operators can fall within three
distinct areas:

  * unquoted and callable, such as, `+`, `-`, and
    most operators

  * unquoted and not callable, such as, `.`, `..`,
    and `..//`, which would be ambiguous when used
    as `Module.OP()`

  * quoted but callable, such as `::`, which is
    ambiguous in the atom syntax but not as calls

This information still remains private, especially
because most times operators are used, we often still
want to learn about their precedence too.