Allow bypassing application mode validation in release spec
Today, there is a mode validation check when doing a Mix Release that
prevents a parent application that has an application mode of
`:permanent`, for example, while a child application has mode `:load`,
as it might be unsafe.

However, some complex applications may need more control over the
application load/start order.  For such cases, the user would need a
way to tell Mix.Release to don't be strict while constructing the
`.rel` file.

Here, we add an option `:validate_application_mode?` to the release
options (default: `true`) which disables such strict check if set to
`false`.  The default value ensures the current behavior is preserved.