expand Kernel.in/2 docs re range check efficiency
Current docs treat membership tests on lists and ranges using
`Kernel.in/2` with the same warning in the doc saying that comparison on
large lists/ranges is inefficient.

This is true for lists, but not for ranges. A more efficient check is
done there which can run in constant time.

Update docs to explain that.

Have omitted going into detail on how membership checks for ranges with
steps are done, as that seems less relevant, and the general comment
that the check can be done efficiently still holds true.