Fix coverage threshold ignored for Total line in Summary
An incorrect variable is being passed to the `display/2` function on line 353 for the Total line in the coverage summary report. The result is that that Total line will always display in red no matter what the threshold is set to. 

The correct threshold value is retrieved on line 348 from the `opts` or the default and is used on line 351 for the body of the summary report. The final call to display then reverts to using the `opts` variable which is incorrect. 

I have been unable to capture the color of text in the summary report in a test, so have been unable to write a test for this change.