Add binary_slice/2 and binary_slice/3
So far Elixir only exposed binary_part/3 as part of its
functions to manipulate binaries, since that's allowed in
guards.

However, `binary_part/3` is quite limited in that it
expects both `start` and `start+size` to fall within the
binary. This could lead developers to use String.slice/2
or String.slice/3 when a binary operation would be more
performant or appropriate.

This pull request therefore adds `binary_slice/2` and
`binary_slice/3` that nicely integrates with binaries
and ranges, and behaves similarly to their `Enum.slice`
and `String.slice` counterparts.