Support disabling BEAM compression in mix releases
In releases, the final step for stripping chunks of out BEAM files is to
gzip compress the file. This change lets users pass `compress: false` to
turn this off. Somewhat unintuitively, turning compression off here
enables better compression at a later step. This is due to the archive
compressors being able to process all BEAM files rather than restarting
compression for each individual BEAM file.

For some Nerves devices, reducing the total number of bytes sent to
update software is much more important than reducing the on-disk usage.
This is due to their network connections being metered. Given enough
devices, even small size redunctions can result in meaningful savings.

Here are more details:

The default Nerves configuration puts all of the BEAM files in one
SquashFS archive. While Nerves also uses gzip for the SquashFS archive
and SquashFS (mostly) individually compresses files as well, letting
SquashFS perform the compression results in 2.5% file size reduction.
This is assumed to be due to using a higher compression level with
building the SquashFS archive. As a side benefit, moving gzip
compression to SquashFS removes the decompression steps from the BEAM
load and it resulted in a ~500ms boot time improvement on GRiSP 2
hardware for a small Nerves app. Presumably the Linux kernel's gzip
decompression is faster than Erlang's or removing the Erlang gunzip
calls completely added up.

When it's possible to compress all BEAM files together (for example,
when making tar files), not compressing BEAM files also helps. This is
because the BEAM files have a lot of similar strings that are now
visible to the archive compressor. It's also possible to use better
compression methods than gzip.

A similar use is the way that delta firmware updates are handled with
Nerves. Delta firmware updates are a way to send down the difference
between the firmware image on a device and the firmware image you want.
These are generated across the full image and should benefit from being
able to work off uncompressed .beam files.