Run ExUnit even when no tests need to be ran
This ensures that we properly call `after_suite` callbacks that were
setup in the `test_helper.exs` or other places. Without running those,
it could leave the app in a bad state. This is especially important for
umbrella testing with partitions as those may do some setup that breaks
the next app setup if not cleaned up.