Make individual files in `escript.build` example better distinguishable
Previously it was not clear to everyone that the escripts entrypoint
can't be in the `mix.exs` file but has to be separate.

This resulted in rare cases where people asked in the elixir forum why
their escripts module wasn't found.

The proposed changes should make this a bit more clear, though could of
course be improved by some explaining words that accompany the 2 snippets.

Fixes #11770