Improve warning message on quoted atoms/calls/keywords
The warning messages stated that an atom containing starting with a
number and containing only ASCII letters, numbers and underscores does
not require quoting, e.g. `:401`, but this is not the case; `:401` is a
syntax error, as is `:1a`, etc.

This commit changes the warning message to make clear that unquoted
atoms/calls and keywords must not start with a number.