Allow any valid name for PartitionSupervisor
The PartitionSupervisor used a named ETS table to manage partitions and as a
result it required the name to be an atom. Only using atom names is
limiting because it prevents the use of global or via tuples for
registration.

This removes the `:named_table` option from a PartitionSupervisor's ETS
table and instead stores a reference to the table with
`:persistent_term`. Without the named table requirement it's possible to
register with any viable supervisor name.