ExUnit diff algebra improved pinned struct type support
Ensure that the diff algebra generation process can handle a case where
the left element has a pinned struct type and the right element is not a
struct. In this case render the pin info and display the pin variable
in the test output.

This prevents ExUnit's diff process from raising for asserts like:
```
assert %^type{} = nil
assert %^type{} = :not_a_struct
```