update DynSup reference in supervisor docs
Follow up on the improvements of https://github.com/elixir-lang/elixir/pull/11970. WDYT?
The idea is to make clear that `DynamicSupervisor` is an efficient way of starting children dynamically (and not the only way).

PS.: @josevalim I wonder if this distinction should be made in the guides as well, what is your opinion about it? I remember reading it and being confused about this particular subject.