Add new.task mix task to generate a new task file
Historically, I haven't used mix tasks as often as I probably might
have because I always needed to look up the docs to remind myself of
the convention for the file path and module name.

This adds a mix task that will generate a new task file in the
proper location.