Remove head blocking on Task.yield_many/2
`yield_many/2` would previously block waiting for
the reply from the first task, then the second task,
and so on.

However, if the first task was slow, it means the
message inbox would grow quickly, only to be processed
later.

This commit makes it so messages are consumed as they
arrive. This yields better performance but uses more
memory as all results have to be accumulated first.
Then, finally, there is one last pass on the data to
associate results to tasks. Given this additional pass,
we now also support killing or ignoring tasks directly
from Task.yield_many/2.