Update docs on how String.length
Elixir's `String.length/1` function does what's expected for UTF8 strings by counting graphemes. However, this correct behavior might deceive people in some cases¹. Check this issue for a good example: https://github.com/elixir-lang/elixir/issues/12154#issuecomment-1258373007. 

Worse is if you use tools online for a sanity check and the languages in question actually count towards the actual string size. This is easier to notice with short words, but not if you are diffing huge chunks of text. 

I've added a section to bring some of the excellent explanations provided into the linked issue to the official docs. I think this is an honest mistake that might be more common than we expect (mainly for people in Latin America) and it would be great to have this little "gotcha" highlighted somewhere in the docs.

PS.: Feel free to make suggestions or move the section somewhere else. The specific problem I want to solve is to make clear that the actual size of the string might be different than you expect from other languages (ie: trying to store 255 chars from `String.slice(text, 0..254)` into a `varchar(255)` column).