Clarification to convert a string to module name by `to_existing_atom`
Every time I try to convert a string to a module name, I receive an error, which I find to be very confusing. I believe the example in the document can help to address this issue. Therefore, I am aware that the document needs to be very simple and basic. You get to decide. 🙏🏻
Hopefully I didn't squander your time.