Consider mix alises in umbrella projects (#12576)
Previously, when an alias existed within an umbrella child, running this
from the umbrella root would result in an error that the mix task was
not found.

This commit checks the aliases of the children, and if any of the
children have the alias then it will be executed in the children. Any
children that do not have the alias will be skipped.