Define "nutrition facts" for "use SomeModule"
One common concern about using Elixir libraries
is that "use SomeModule" does not make clear how
it impacts the caller. Most of the time those
changes are minor but it is impossible to know
without traversing the source code.

This pull request suggests defining a summary
alongside each module that succintly explains
the impact of using it.

This is documented as best practice in our guides
and we added the note to all of Elixir modules.