Limit number of structs in iex autocomplete
Getting structs for iex autocomplete is expensive: modules need
to be loaded and checked to see if they define structs. This means
that autocompleting for an empty struct name or a short one can be
very slow in codebases with many modules and dependencies. In my
work codebase, it might take several seconds for the autocomplete
to return something and iex get stuck meanwhile.

This PR limits the number of structs, so that if a query that would
return too many results is made, iex returns the completion quickly
enough.