Add --no-runtime-exs
This adds a new CLI option to `mix run` that will skip loading the
runtime.exs file.

This is useful when you want to load a VM with all of the project's code
and dependencies, but don't want to actually start it in any way.

runtime.exs can use your project's dependencies, which might not be
compiled yet if also using the `--no-compile` option.

Normally you can simply gate your entire runtime.exs file in a condition
that checks for an environment variable, but in the case of developer
tooling like a language server, you can't alter the code.