Improve UndefinedFunctionError for mispelled module
This is a proof of concept to see if we can improve error messages when
a module name is mispelled or otherwise incorrect. This can happen when
a user misspells a module, or does not fully qualify it. Ideally, the
elixir compiler can provide suggestions for what they meant.

If we were to merge a version of this PR, we would get error messages like

```elixir
iex(1)> Enu.map([1, 2, 3], & &1 + 1)
** (UndefinedFunctionError) function Enu.map/2 is undefined (module Enu is not available). Did you mean:

      * Enum.map/2

    Enu.map([1, 2, 3], #Function<42.125776118/1 in :erl_eval.expr/6>)
    iex:1: (file)
```

```elixir
iex(1)> defmodule Namespace.SomeModule do
...(1)>   def foo, do: :bar
...(1)> end
{:module, Namespace.SomeModule,
 <<70, 79, 82, 49, 0, 0, 5, 36, 66, 69, 65, 77, 65, 116, 85, 56, 0, 0, 0, 186,
   0, 0, 0, 18, 27, 69, 108, 105, 120, 105, 114, 46, 78, 97, 109, 101, 115, 112,
   97, 99, 101, 46, 83, 111, 109, 101, 77, ...>>, {:foo, 0}}
iex(2)> SomeModule.foo()
** (UndefinedFunctionError) function SomeModule.foo/0 is undefined (module SomeModule is not available). Did you mean:

      * Namespace.SomeModule.foo/0

    SomeModule.foo()
    iex:2: (file)
```

I saw two experienced developers who are new two Elixir get stuck on
this issue on the same day, not including a mention on Twitter!

[Twitter Conversation](https://twitter.com/davydog187/status/1687628454240428034?s=20)

- [ ] Better heuristic that is accurate for most usecases
- [ ] Consider other edge cases and measure results (forgetten alias?)
- [ ] Possibly turn off
outside of dev?
- [ ] Remove `to_atom/1`
- [ ] Write more tests

* https://groups.google.com/g/elixir-lang-core/c/-IHneszMheI/m/cCCxHB9PBwAJ
* https://github.com/elixir-lang/elixir/pull/5665