Consistently raise on invalid ANSI sequences in `IO.ANSI.format` (#12842)
If `enabled` is set to `true` an invalid ANSI sequence will raise
an `ArgumentError`. On the other hand if `enabled` is set to `false`
no error is raised and the sequence is silently ignored.

This can lead to inconsistent behaviours for CLI tools based on a
configuration setting. Additionally since in tests a common pattern
is to disable ANSI sequences you may have successful tests but
exceptions when executing the code in production with enabled ANSI
sequences.