Point out dangers around `Task` and a lot of data
The `Task` module is one of the coolest modules in elixir and
is probably the first contact and experience of a lot of
beginners with parallelism in elixir. I hence find it worthwhile
to warn about the memory copying and its impacts here as it might
easily lead to unwelcome results, so it's worth pointing out.

This is only a first draft, not sure if this is the best place
to put it - feedback very welcome.

Honestly `Task.async`/`Task.await` may be too magical/too much of
a nice high level magical abstraction. I'm somewhat ashamed to
admit that while I knew that sending a message would copy the
data I never made the connection to `Task.async` myself and I
had an "oh shit" moment just yesterday :sweat-smile: Which,
totally should have occurred to me before as I knew they were
processes and wouldn't magically change the rules on how
processes work. I also asked some elixir friends and saw similiar
reactions in some of them.

I think it's also fair to assume that the BEAM wouldn't need to
copy immutable data (cos it's immutable anyhow), I know why it
does but it's not obvious. We might want to/need to work on making
it more obvious in other parts of the documentation as well but
I thought this was the best starting point.