Mention the fixed regression
Kudos go to @sabiwara, full log/discussion at:
https://github.com/bencheeorg/benchee/issues/406

As best as I learned this was introduced in #11420 and fixed in
aabe465
It shouldn't affect almost anyone except for scripting usage of
elixir and (probably most notably) benchee benchmarks that don't
call functions in a module.

I think this warrants a Changelog entry. Not sure if [Compiler] is
right and my entry is probably way too wordy - happy to adjust it.

I'll also write up a notice at benchee and potentially a blog post
so that people can know to potentially redo benchmarks they ran
on 1.14-1.15 by either wrapping the functions or the
`Benchee.run` invocation in modules or running on an unaffected
Elixir version.