Pull requests need approval of one committer?
Browsing through the contributing docs earlier I noticed it
mentioned needing approval from 2 committers. However, recent
changes I've seen were fine to go from 1 committer so I assume
the guidelines is outdated and updated it.

Of course, if that's only temporary or has some more nuance
happy for this to be closed :grin: