Add atom singleton types to descr
The :atom field now contains a representation for all possible
atom types. The representation is a pair `{s, bool}` where `s`
is a set of atoms. If `bool` is `true` it represents the union
of the atoms in `s`. Else, it represents every atom except those
in `s`.

`boolean()` is defined as the singletons true, false.